requirements.txt
accelerate==0.33.0
certifi==2024.7.4
charset-normalizer==3.3.2
cmake==3.30.1
diffusers==0.28.0
einops==0.8.0
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.3
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
Jinja2==3.1.4
lazy_loader==0.4
lit==18.1.8
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
psutil==6.0.0
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
sympy==1.13.1
tifffile==2024.7.24
timm==0.6.7
tokenizers==0.15.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.66.4
transformers==4.36.2
triton==2.0.0
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
